Update clm.rst
This part is made by copying and pasting from clm5.0 user's guide by NCAR. (https://escomp.github.io/ctsm-docs/versions/release-clm5.0/html/users_guide/setting-up-and-running-a-case/choosing-a-compset.html#). I am not sure if it will cause any copyright issues. But for land-only experiments, the CLM in the NorESM is no different from the CLM in the CESM.