Update lnd_model.rst
I have added some information about the biogeophysical and biogeochemical processes that CLM5 takes care of. Because of the nature of land surface modeling, the processes are so many on different spatial scales (from leave stomata to global carbon cycle), and all of which are important and nonnegligible (for clm5_bgc_crop, nearly all the processes above are included in the simulation). I, therefore, think it is probably best to add a list of simulated processes without detailed information on each one of them. This list is directly from the CLM5 technical note, so I add a weblink so that people can get help from it if they need it.