Provide more detailed example for BLOM user_nl_blom
The build system seem not to handle lower case written parameters for BLOM - so, in both, the ./components/blom/cime_config/buildnml and user_nl_blom, parameters need to be spelled capital. Further, no quotation marks are required when providing a vector as input - so I added  this as an additional example.